Add ClrMethodData
