+++ CilTools.Runtime/ClrConstructorInfo.cs
@@ -7,8 +7,8 @@ using System.Globalization;
 using System.Text;
 using System.Reflection;
 using Microsoft.Diagnostics.Runtime;
-using CilTools.BytecodeAnalysis;
 using CilTools.Reflection;
+using CilTools.Runtime.Methods;
 
 namespace CilTools.Runtime
 {
@@ -17,28 +17,17 @@ namespace CilTools.Runtime
     /// </summary>
     public class ClrConstructorInfo : ConstructorInfo, ICustomMethod
     {
-        ClrMethod method;
-        ClrAssemblyInfo assembly;
-        DataTarget target;
-        ClrTypeInfo type;
-
-        //backing dynamic method, if this is a method from dynamic module
-        ClrDynamicMethod dynamicMethod = null;
-        bool dynamicMethodInitialized = false;
+        ClrMethodData md;
 
         internal ClrConstructorInfo(ClrMethod m, ClrTypeInfo owner)
         {
-            this.method = m;
-            this.assembly = (ClrAssemblyInfo)owner.Assembly;
-            this.type = owner;
-
-            if (assembly != null) this.target = assembly.InnerModule.Runtime.DataTarget;
+            this.md = new ClrMethodData(m, owner);
         }
 
         /// <summary>
         /// Gets the underlying ClrMD method object
         /// </summary>
-        public ClrMethod InnerMethod { get { return this.method; } }
+        public ClrMethod InnerMethod { get { return this.md.InnerMethod; } }
 
         /// <inheritdoc/>
         public Type ReturnType
@@ -49,67 +38,13 @@ namespace CilTools.Runtime
         /// <inheritdoc/>
         public ITokenResolver TokenResolver
         {
-            get { return this.assembly; }
+            get { return this.md.Assembly; }
         }
-
-        ClrDynamicMethod FindDynamicMethod()
-        {
-            if (this.method.Type == null) return null;
-            ClrModule module = this.method.Type.Module;
-            if (module == null) return null;
-
-            if (this.dynamicMethodInitialized)
-            {
-                return this.dynamicMethod;
-            }
-
-            //try to lookup dynamic method that backs up this 
-            //method in dynamic module
-
-            int token = 0;
-
-            unchecked { token = (int)this.method.MetadataToken; }
-
-            ClrDynamicMethod ret = this.assembly.AssemblyReader.GetDynamicAssemblyMethod(
-                module.Address,
-                token
-                );
-
-            this.dynamicMethod = ret;
-            this.dynamicMethodInitialized = true;
-            return ret;
-        }
-
+        
         /// <inheritdoc/>
         public byte[] GetBytecode()
         {
-            byte[] il;
-            int bytesread;
-            ILInfo ildata = method.IL;
-
-            if (ildata == null)
-            {
-                //P/Invoke methods does not have IL body
-                if (this.method.IsPInvoke) return new byte[0];
-
-                //try to lookup dynamic method that backs up this 
-                //method in dynamic module
-                ClrDynamicMethod dm = this.FindDynamicMethod();
-
-                if (dm != null)
-                {
-                    il = dm.GetBytecode();
-                    if (il != null) return il;
-                }
-
-                throw new CilParserException("Cannot read IL of the method " + method.Name);
-            }
-            else
-            {
-                il = new byte[ildata.Length];
-                target.ReadProcessMemory(ildata.Address, il, ildata.Length, out bytesread);
-                return il;
-            }
+            return this.md.GetBytecode();
         }
 
         /// <inheritdoc/>
@@ -139,42 +74,13 @@ namespace CilTools.Runtime
         /// <inheritdoc/>
         public ExceptionBlock[] GetExceptionBlocks()
         {
-            //P/Invoke methods does not have exception blocks
-            if (this.method.IsPInvoke) return new ExceptionBlock[] { };
-
-            //try to lookup dynamic method that backs up this 
-            //method in dynamic module
-            ClrDynamicMethod dm = this.FindDynamicMethod();
-
-            if (dm != null)
-            {
-                return dm.GetExceptionBlocks();
-            }
-            else
-            {
-                return new ExceptionBlock[] { };
-            }
+            return this.md.GetExceptionBlocks();
         }
 
         /// <inheritdoc/>
         public override MethodAttributes Attributes
         {
-            get
-            {
-                MethodAttributes ret = (MethodAttributes)0;
-                if (method.IsAbstract) ret |= MethodAttributes.Abstract;
-                if (method.IsFinal) ret |= MethodAttributes.Final;
-                if (method.IsInternal) ret |= MethodAttributes.Assembly;
-                if (method.IsPrivate) ret |= MethodAttributes.Private;
-                if (method.IsProtected) ret |= MethodAttributes.Family;
-                if (method.IsPublic) ret |= MethodAttributes.Public;
-                if (method.IsStatic) ret |= MethodAttributes.Static;
-                if (method.IsVirtual) ret |= MethodAttributes.Virtual;
-                if (method.IsPInvoke) ret |= MethodAttributes.PinvokeImpl;
-                if (method.IsSpecialName) ret |= MethodAttributes.SpecialName;
-                if (method.IsRTSpecialName) ret |= MethodAttributes.RTSpecialName;
-                return ret;
-            }
+            get { return this.md.GetAttributes(); }
         }
 
         /// <inheritdoc/>
@@ -191,7 +97,7 @@ namespace CilTools.Runtime
 
         /// <inheritdoc/>
         public override object Invoke(object obj, BindingFlags invokeAttr, Binder binder, object[] parameters,
-            System.Globalization.CultureInfo culture)
+            CultureInfo culture)
         {
             throw new InvalidOperationException("Cannot invoke methods on type loaded into reflection-only context");
         }
@@ -211,7 +117,7 @@ namespace CilTools.Runtime
         /// <inheritdoc/>
         public override Type DeclaringType
         {
-            get { return this.type; }
+            get { return this.md.OwnerType; }
         }
 
         /// <inheritdoc/>
@@ -259,7 +165,7 @@ namespace CilTools.Runtime
         /// <inheritdoc/>
         public override string Name
         {
-            get { return method.Name; }
+            get { return this.md.InnerMethod.Name; }
         }
 
         /// <inheritdoc/>
@@ -273,7 +179,7 @@ namespace CilTools.Runtime
         {
             get
             {
-                return (int)method.MetadataToken;
+                return (int)this.md.InnerMethod.MetadataToken;
             }
         }
